refactor(engine): relocate worker.py to src/tuners/engine/
Move Worker (per-instance configuration + performance state) into the shared
execution engine. It depends only on src.{knobs,utils,config} — no tuning
strategy — and is consumed by every tuner (PBT, BO, LHS), the orchestrator, and
evolution. It belongs beside barriers/restart_policy in src/tuners/engine/.

- git mv worker.py -> src/tuners/engine/ (history preserved)
- add Worker to engine/__init__.py public surface
- repoint all 15 importers (core, evolution, population, main, orchestrator,
  lhs_design, bo_baseline, tests)
- src/tuner/core/__init__.py keeps re-exporting Worker for the transitional
  import surface

Part of ADR-006 tuners/ unification. Relocate-first; no behavior change.
