+++ src/scripts/bo_baseline/cotenant.py
@@ -54,7 +54,7 @@ from typing import List, Optional
 from src.config.database import DatabaseConfig
 from src.knobs.knob_space import KnobSpace
 from src.tuners.engine.barriers import GenerationBarrier
-from src.tuner.core.worker import Worker
+from src.tuners.engine.worker import Worker
 from src.tuner.benchmark.orchestrator import WorkloadOrchestrator
 from src.utils.environments.base import DatabaseEnvironment
 from src.utils.logger import get_logger
