+++ src/scripts/bo_baseline/objective.py
@@ -5,7 +5,7 @@ from typing import Dict, Optional, Tuple, TYPE_CHECKING
 from ConfigSpace import Configuration
 
 from src.knobs.knob_space import KnobSpace
-from src.tuner.core.worker import Worker
+from src.tuners.engine.worker import Worker
 from src.tuners.engine.barriers import GenerationBarrier
 from src.tuner.benchmark.orchestrator import WorkloadOrchestrator
 from src.utils.metrics import PerformanceMetrics
