+++ src/scripts/bo_baseline/runner.py
@@ -17,7 +17,7 @@ from smac.runhistory.dataclasses import TrialInfo, TrialValue
 from smac.runhistory.enumerations import StatusType
 
 from src.knobs import get_knob_space
-from src.tuner.core.worker import Worker
+from src.tuners.engine.worker import Worker
 from src.tuner.benchmark.orchestrator import (
     WorkloadOrchestrator,
     WorkloadOrchestratorConfig,
