+++ src/tuners/engine/__init__.py
@@ -4,6 +4,7 @@ Tuner Execution Engine
 
 Tuner-agnostic execution primitives shared across all tuning strategies:
 
+- Worker: per-instance configuration + performance state
 - Lockstep generation barriers (B1–B17 synchronisation)
 - TuningMode-driven restart policy
 
@@ -15,9 +16,11 @@ dependency graph.
 
 from src.tuners.engine.barriers import GenerationBarrier, BARRIER_NAMES
 from src.tuners.engine.restart_policy import should_restart
+from src.tuners.engine.worker import Worker
 
 __all__ = [
     "GenerationBarrier",
     "BARRIER_NAMES",
     "should_restart",
+    "Worker",
 ]
